GameObjectTypes
moved unnecessary elements cluttering the top of main into a separate header file, these can either be left alone or integrated into existing classes
